frontend/requirements.txt
streamlit
streamlit[auth]
groq
markdown
scikit-learn
langchain
langchain-text-splitters

python-magic

unidecode

psycopg

requests
atlassian-python-api

matplotlib
tqdm

weaviate-client